% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBtRED3.R
\name{PBtRED3}
\alias{PBtRED3}
\title{Partially Balanced Ternary Residual Effect Designs for Number of Treatments, v ( >= 5)}
\usage{
PBtRED3(v)
}
\arguments{
\item{v}{Number of treatments (>= 5)}
}
\value{
It returns a new class of PBTREDs along with its parameters, Information matrix (C), Average Variance Factor (AVF), and Canonical Efficiency Factor (CEF) for both treatment and residual effects.
}
\description{
This function generates a new class of partially balanced ternary residual effect designs (PBTREDs) for all treatments, v (>= 5) in periods, p = (v+3)/2 if v is odd, otherwise p = (v+2)/2, and the number of sequences, n = 2v.
}
\examples{
library(TREDesigns)
PBtRED3(v = 5)
}
