% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBtRED4.R
\name{PBtRED4}
\alias{PBtRED4}
\title{Partially Balanced Ternary Residual Effect Designs for Number of Treatments, v = 2m (m >= 4 & even)}
\usage{
PBtRED4(m)
}
\arguments{
\item{m}{Any even number (>= 4)}
}
\value{
It returns a new class of PBTREDs along with its parameters, Information matrix (C), Average Variance Factor (AVF), and Canonical Efficiency Factor (CEF) for both treatment and residual effects.
}
\description{
This function generates a new class of partially balanced ternary residual effect designs (PBTREDs) for the number of treatments, v = 2m; where m >= 4 & even. The number of periods, p = (v+4)/2, while the number of sequences, n = v.
}
\examples{
library(TREDesigns)
PBtRED4(m = 4)
}
