% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_RED.R
\name{Study_RED}
\alias{Study_RED}
\title{Studying Properties of Ternary Residual Effect Designs}
\usage{
Study_RED(design)
}
\arguments{
\item{design}{Provide a ternary residual effect design}
}
\value{
It returns Information matrix (C), Average Variance Factor (AVF), and Canonical Efficiency Factor (CEF) for both treatment and residual effects for a given ternary residual design.
}
\description{
To study the properties of any given ternary residual effect design.
}
\examples{
library(TREDesigns)
design=PBtRED3(v = 5)$PBTRED
Study_RED(design)
}
